# NIFI-3236 - SplitJson performance improvements